+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2022-10-30
+
+- Add `IsFriend` to user details
+- Add ability to get friends of the authenticated user
+- Add ability to add/remove another user as a friend of the authenticated user
+
 ## [2.1.0] - 2022-10-30
 
 - Add ability to edit submission/comment text
@@ -75,6 +81,7 @@ This release also includes a number of new tests, both for new and existing func
   - Submissions
   - Comments
 
+[2.2.0]: https://github.com/JedS6391/Reddit.NET/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/JedS6391/Reddit.NET/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/JedS6391/Reddit.NET/compare/1.3.0...2.0.0
 [1.3.0]: https://github.com/JedS6391/Reddit.NET/compare/1.2.0...1.3.0